Refactor: Split classes up into files
In preparation for testing with JUnit
