+++ src/main/java/duke/DukeException.java
@@ -1,0 +1,8 @@
+package duke;
+
+class DukeException extends Exception {
+
+  public DukeException(String message) {
+    super(message);
+  }
+}
