+++ src/test/java/duke/DukeTest.java
@@ -1,0 +1,7 @@
+package duke;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+import org.junit.jupiter.api.Test;
+
+public class DukeTest {}
